--- fmriprep_denoise/metrics/__init__.py
@@ -1,7 +1,0 @@
-from .quality_control_connectivity import qcfc, partial_correlation, fdr, calculate_median_absolute
-from .distance_dependency import get_atlas_pairwise_distance
-from .network_modularity import louvain_modularity
-
-__all__ = ['qcfc', 'fdr', 'calculate_median_absolute',
-           'get_atlas_pairwise_distance', 'louvain_modularity',
-           'partial_correlation']
