--- fmriprep_denoise/utils/data/custome_templateflow/dataset_description.json
@@ -1,34 +1,0 @@
-{
-  "BIDSVersion": "1.6.0",
-  "Name": "Collection of atlases reorganised by TemplateFlow convention.",
-  "DatasetType": "derivative",
-  "GeneratedBy": [
-    {
-      "Name": "difumo_segmentation",
-      "Version": "0.1.0",
-      "CodeURL": "https://github.com/SIMEXP/difumo_segmentation"
-    },
-    {
-      "Name": "mist2templateflow",
-      "Version": "0.1.0",
-      "CodeURL": "https://github.com/SIMEXP/mist2templateflow"
-    }
-  ],
-  "SourceDatasets": [
-    {
-      "DOI": "10.1016/j.neuroimage.2020.117126.",
-      "URL": "https://osf.io/k8w5s/",
-      "Version": "1"
-    },
-    {
-      "DOI": "10.1016/j.neuroimage.2020.117126.",
-      "URL": "https://doi.org/10.12688/mniopenres.12767.2",
-      "Version": "2"
-    },
-    {
-      "DOI": "10.1101/2021.02.10.430678",
-      "URL": "https://www.templateflow.org/",
-      "Version": "1.4.1"
-    }
-  ]
-}
