--- fmriprep_denoise/utils/data/custome_templateflow/tpl-MNI152NLin2009bSym/tpl-MNI152NLin2009bSym_res-03_atlas-MIST_desc-12_dseg.tsv
@@ -1,13 +1,0 @@
-roi	label	name	size	symmetry	laterality	x	y	z	neighbour	parent	overlap
-1	BG_THAL	BASAL_GANGLIA_and_THALAMUS	2476	0.92	0.0	-0.43	-2.65	5.04	[2, 3, 8, 11]	7	1.0
-2	AUDnet_PINS	AUDITORY_NETWORK_and_POSTERIOR_INSULA	2947	0.86	0.031	-2.37	-17.81	8.52	[1, 3, 4, 7, 10, 11]	7	1.0
-3	LIMnet	MESOLIMBIC_NETWORK	5556	0.87	0.013	-0.6	1.39	-21.82	[1, 2, 4, 5, 6, 8, 9, 10, 11, 12]	2	1.0
-4	DMnet_l	DEFAULT_MODE_NETWORK_lateral	3688	0.78	0.043	-1.81	-20.8	-1.27	[2, 3, 8, 9, 10, 11]	5	0.9303145336225596
-5	VISnet	VISUAL_NETWORK	6347	0.89	-0.0559999999999999	1.58	-79.25	5.54	[3, 6, 9, 12]	4	1.0
-6	DMnet_pm	DEFAULT_MODE_NETWORK_posteromedial	2002	0.88	0.038	-0.73	-55.67	28.55	[3, 5, 7, 9, 11]	5	0.8096903096903096
-7	MOTnet	SOMATOMOTOR_NETWORK	4738	0.89	0.013	-0.35	-20.69	53.91	[2, 6, 9, 10, 11]	3	1.0
-8	DMnet_am_lhAG	DEFAULT_MODE_NETWORK_anteromedial_and_left_ANGULAR_GYRUS	4030	0.79	0.135	-5.07	38.08	21.8	[1, 3, 4, 9, 10, 11]	5	0.9761786600496278
-9	VVIS_DVIS	VENTRAL_VISUAL_STREAM_and_DORSAL_VISUAL_STREAM	4744	0.79	0.043	-2.11	-59.96	21.08	[3, 4, 5, 6, 7, 8, 10, 11, 12]	6	0.9911467116357504
-10	FPnet	FRONTO_PARIETAL_NETWORK	6275	0.73	-0.095	3.92	13.95	24.93	[2, 3, 4, 7, 8, 9, 11]	6	0.9832669322709164
-11	VATTnet_SALnet	VENTRAL_ATTENTION_NETWORK_and_SALIENCE_NETWORK	7759	0.77	-0.011	0.55	-2.13	37.1	[1, 2, 3, 4, 6, 7, 8, 9, 10]	7	0.8735661812089187
-12	CER	CEREBELLUM	7228	0.96	0.0139999999999999	-0.41	-64.36	-35.89	[3, 5, 9]	1	1.0
