--- fmriprep_denoise/utils/data/custome_templateflow/tpl-MNI152NLin2009bSym/tpl-MNI152NLin2009bSym_res-03_atlas-MIST_desc-20_dseg.tsv
@@ -1,21 +1,0 @@
-roi	label	name	size	symmetry	laterality	x	y	z	neighbour	parent	overlap
-1	BG_THAL	BASAL_GANGLIA_and_THALAMUS	2476	0.92	0.0	-0.43	-2.65	5.04	[3, 5, 8, 10, 19]	1	1.0
-2	MOTnet_v	SOMATOMOTOR_NETWORK_ventral	913	0.78	0.025	-2.15	-5.03	28.29	[7, 8, 11, 12, 13, 19]	7	1.0
-3	ORBcor_NACC	ORBITOFRONTAL_CORTEX_and_NUCLEUS_ACCUMBENS	1411	0.81	0.06	-1.91	31.59	-17.84	[1, 5, 10, 18, 19, 20]	3	1.0
-4	ASCER	ASSOCIATIVE_CEREBELLUM	3445	0.87	0.057	-1.57	-54.11	-35.64	[9, 10, 15, 16, 17]	12	1.0
-5	PGACcor_VMPFcor	PERIGENUAL_ANTERIOR_CINGULATE_CORTEX_and_VENTROMEDIAL_PREFRONTAL_CORTEX	1335	0.92	0.034	-0.49	45.41	0.63	[1, 3, 13, 20]	8	1.0
-6	ITgyr_Tpol	INFERIOR_TEMPORAL_GYRUS_and_TEMPORAL_POLE	1700	0.85	0.016	-0.31	-7.33	-33.64	[10, 16, 18]	3	1.0
-7	FPTCnet	FRONTO_PARIETAL_TASK_CONTROL_NETWORK	2442	0.72	-0.118	4.97	9.36	26.13	[2, 8, 11, 12, 13, 16, 18, 19, 20]	10	1.0
-8	AUDnet_PINS	AUDITORY_NETWORK_and_POSTERIOR_INSULA	2575	0.86	0.05	-3.23	-19.86	10.46	[1, 2, 7, 10, 13, 18, 19, 20]	2	0.9813592233009708
-9	MVISnet	MEDIAL_VISUAL_NETWORK	2694	0.84	0.023	-0.23	-74.9	12.55	[4, 10, 14, 15, 16, 17]	5	1.0
-10	AMY_HIPP_Pisul	AMYGDALA_and_HIPPOCAMPUS_and_Peri_Insular_Sulcus	2445	0.86	-0.016	-0.05	-9.97	-15.9	[1, 3, 4, 6, 8, 9, 14, 16, 18, 19, 20]	3	1.0
-11	MOTnet_d	SOMATOMOTOR_NETWORK_dorsal	3733	0.88	-0.017	0.74	-24.23	60.06	[2, 7, 12, 14, 16, 19, 20]	7	0.988748995446022
-12	VATTnet_m	VENTRAL_ATTENTION_NETWORK_medial	3294	0.83	0.036	-1.34	-4.82	51.29	[2, 7, 11, 13, 14, 16, 20]	11	0.9565877352762598
-13	DMnet_l	DEFAULT_MODE_NETWORK_lateral	3797	0.55	0.367	-14.51	20.48	34.88	[2, 5, 7, 8, 12, 16, 18, 19, 20]	8	0.7097708717408481
-14	DMnet_pm	DEFAULT_MODE_NETWORK_posteromedial	2414	0.9	0.008	-0.47	-56.87	32.37	[9, 10, 11, 12, 16]	6	0.8293289146644574
-15	LVISnet	LATERAL_VISUAL_NETWORK	3549	0.79	-0.147	3.81	-83.32	0.58	[4, 9, 16, 17]	5	1.0
-16	VVIS_DVIS	VENTRAL_VISUAL_STREAM_and_DORSAL_VISUAL_STREAM	5098	0.78	0.07	-1.95	-58.61	21.54	[4, 6, 7, 9, 10, 11, 12, 13, 14, 15, 17, 18, 19, 20]	9	0.8415064731267163
-17	MOCER	MOTOR_CEREBELLUM	3783	0.87	-0.025	0.65	-73.7	-36.11	[4, 9, 15, 16]	12	1.0
-18	DMnet_l	DEFAULT_MODE_NETWORK_lateral	3268	0.79	0.06	-2.94	-14.68	-4.6	[3, 6, 7, 8, 10, 13, 16, 19, 20]	4	0.9684822521419828
-19	VATTnet_l	VENTRAL_ATTENTION_NETWORK_lateral	3206	0.73	-0.049	2.2	-8.57	17.62	[1, 2, 3, 7, 8, 10, 11, 13, 16, 18, 20]	11	0.868995633187773
-20	FPnet	FRONTO_PARIETAL_NETWORK	4212	0.55	-0.29	11.4	25.67	21.84	[3, 5, 7, 8, 10, 11, 12, 13, 16, 18, 19]	10	0.6711775878442545
