--- fmriprep_denoise/utils/data/custome_templateflow/tpl-MNI152NLin2009bSym/tpl-MNI152NLin2009bSym_res-03_atlas-MIST_desc-36_dseg.tsv
@@ -1,37 +1,0 @@
-roi	label	name	size	symmetry	laterality	x	y	z	neighbour	parent	overlap
-1	CER5	CEREBELLUM_I-V	747	0.91	0.007	-0.4	-46.11	-23.17	[5, 13, 16, 24, 26, 30]	4	1.0
-2	PUT	PUTAMEN	1033	0.88	0.032	-1.18	1.38	1.51	[8, 10, 19, 25, 28, 30]	1	1.0
-3	MOTnet_vl	SOMATOMOTOR_NETWORK_ventrolateral	913	0.78	0.025	-2.15	-5.03	28.29	[9, 21, 25, 28, 33, 35, 36]	2	1.0
-4	PCcor_PRC	POSTERIOR_CINGULATE_CORTEX_and_PRECUNEUS	776	0.81	-0.048	0.7	-48.75	32.07	[14, 22, 26, 32]	14	1.0
-5	CER9	CEREBELLUM_IX	966	0.87	0.062	-0.57	-53.31	-45.54	[1, 6, 13, 20, 30]	4	1.0
-6	CER7b	CEREBELLUM_VIIb	816	0.66	-0.104	-0.1	-71.52	-52.09	[5, 7, 13, 20]	17	1.0
-7	CERCR1	CEREBELLUM_CRUSI	1133	0.76	0.098	-3.75	-79.71	-21.9	[6, 13, 16, 17, 20, 24, 26]	17	1.0
-8	THAL_CAUDN	THALAMUS_and_CAUDATE_NUCLEUS	1443	0.87	-0.023	0.11	-5.54	7.58	[2, 10, 18, 28, 30]	1	1.0
-9	MOTnet_dl	SOMATOMOTOR_NETWORK_dorsolateral	1026	0.65	-0.166	6.8	-17.21	54.68	[3, 15, 21, 23, 33, 35]	11	1.0
-10	ORBcor_NACC	ORBITOFRONTAL_CORTEX_and_NUCLEUS_ACCUMBENS	1411	0.81	0.06	-1.91	31.59	-17.84	[2, 8, 18, 19, 28, 30, 34]	3	1.0
-11	VISnet_md	VISUAL_NETWORK_mediodorsal	1023	0.81	0.012	0.08	-84.0	27.99	[17, 22, 24, 26, 32]	9	1.0
-12	ITgyr_Tpol	INFERIOR_TEMPORAL_GYRUS_and_TEMPORAL_POLE	1700	0.85	0.016	-0.31	-7.33	-33.64	[16, 19, 30, 31]	6	1.0
-13	CER8ab	CEREBELLUM_VI_VIIIab	1649	0.84	0.029	-0.74	-58.19	-34.64	[1, 5, 6, 7, 16, 17, 20, 24, 26, 30]	4	1.0
-14	IMsul	INFERIOR_MARGINAL_SULCUS	1107	0.84	0.062	-0.57	-26.16	42.25	[4, 15, 21, 22, 27, 29, 32, 33]	12	0.8482384824
-15	MOTnet_m	SOMATOMOTOR_NETWORK_medial	1693	0.77	0.08	-2.42	-22.39	63.65	[9, 14, 21, 27, 32, 33]	11	1.0
-16	VVIS	VENTRAL_VISUAL_STREAM	1903	0.74	0.116	-4.26	-52.57	-9.1	[1, 7, 12, 13, 17, 22, 24, 26, 30, 31]	16	1.0
-17	PVISnet	POSTERIOR_VISUAL_NETWORK	1828	0.77	-0.049	0.27	-91.25	0.73	[7, 11, 13, 16, 20, 22, 24, 26]	15	1.0
-18	PGACcor_VMPFcor	PERIGENUAL_ANTERIOR_CINGULATE_CORTEX_and_VENTROMEDIAL_PREFRONTAL_CORTEX	1335	0.92	0.034	-0.49	45.41	0.63	[8, 10, 29, 34, 36]	5	1.0
-19	INS_v_PIsul	INSULA_ventral_and_PERI_INSULAR_SULCUS	1433	0.8	-0.015	-0.49	6.43	-15.28	[2, 10, 12, 25, 28, 30, 31, 34]	10	0.7069085834
-20	CERCR2	CEREBELLUM_CRUSII	1917	0.8	-0.02	1.81	-70.24	-38.42	[5, 6, 7, 13, 17]	17	0.9567031821
-21	PMOTcor_SMOTcor	PREMOTOR_CORTEX_and_SUPPLEMENTARY_MOTOR_CORTEX	2355	0.8	0.011	-1.52	3.58	55.47	[3, 9, 14, 15, 23, 29, 35, 36]	12	1.0
-22	DVIS	DORSAL_VISUAL_STREAM	1683	0.61	-0.175	6.11	-69.37	39.12	[4, 11, 14, 16, 17, 23, 24, 26, 27, 31, 32, 33, 35, 36]	16	1.0
-23	R_DMnet_l	right_DEFAULT_MODE_NETWORK_lateral	1509	0.0	-1.0	40.83	-11.98	41.2	[9, 21, 22, 25, 29, 31, 33, 35, 36]	20	0.9310801856
-24	LVISnet	LATERAL_VISUAL_NETWORK	1721	0.57	-0.25	7.56	-74.9	0.43	[1, 7, 11, 13, 16, 17, 22, 26, 30]	15	1.0
-25	AUDnet_PINS	AUDITORY_NETWORK_and_POSTERIOR_INSULA	2575	0.86	0.05	-3.23	-19.86	10.46	[2, 3, 19, 23, 28, 31, 33, 35, 36]	8	1.0
-26	MVISnet	MEDIAL_VISUAL_NETWORK	1671	0.86	0.031	-0.42	-69.33	3.1	[1, 4, 7, 11, 13, 16, 17, 22, 24, 30, 32]	9	1.0
-27	SPlob	SUPERIOR_PARIETAL_LOBULE	826	0.72	0.075	-2.14	-48.67	66.48	[14, 15, 22, 32, 33, 35]	16	0.6452784504
-28	AINS_d	ANTERIOR_INSULA_dorsal	1317	0.68	-0.13	5.97	15.46	2.37	[2, 3, 8, 10, 19, 25, 29, 31, 34, 35]	19	1.0
-29	EXECnet	EXECUTIVE_CONTROL_NETWORK	1709	0.71	0.176	-5.84	36.36	28.11	[14, 18, 21, 23, 28, 34, 35, 36]	20	0.9198361615
-30	AMY_HIPP	AMYGDALA_and_HIPPOCAMPUS	1858	0.84	0.04	-1.31	-30.1	-12.67	[1, 2, 5, 8, 10, 12, 13, 16, 19, 24, 26, 32]	10	0.7707212056
-31	MTgyr	MIDDLE_TEMPORAL_GYRUS	2504	0.85	-0.014	0.57	-28.08	-5.47	[12, 16, 19, 22, 23, 25, 28, 33, 36]	18	1.0
-32	PRC	PRECUNEUS	1907	0.83	0.048	-1.03	-60.46	37.19	[4, 11, 14, 15, 22, 26, 27, 30, 35]	14	0.8589407446
-33	PCsul_SMgyr	POSTCENTRAL_SULCUS_and_SUPRAMARGINAL_GYRUS	1977	0.75	0.008	-1.1	-30.83	42.54	[3, 9, 14, 15, 22, 23, 25, 27, 31, 35, 36]	19	0.7202832575
-34	VLPFcor	VENTROLATERAL_PREFRONTAL_CORTEX	2161	0.67	0.137	-6.98	44.12	-4.13	[10, 18, 19, 28, 29, 35, 36]	20	0.5714946784
-35	FPTCnet	FRONTO_PARIETAL_TASK_CONTROL_NETWORK	2609	0.63	-0.081	4.87	-2.41	30.18	[3, 9, 21, 22, 23, 25, 27, 28, 29, 32, 33, 34, 36]	7	0.8738980452
-36	DMnet_l	DEFAULT_MODE_NETWORK_lateral	3556	0.5	0.415	-17.22	21.71	34.68	[3, 18, 21, 22, 23, 25, 29, 31, 33, 34, 35]	13	1.0
