--- fmriprep_denoise/utils/data/custome_templateflow/tpl-MNI152NLin2009bSym/tpl-MNI152NLin2009bSym_res-03_atlas-MIST_desc-64_dseg.tsv
@@ -1,65 +1,0 @@
-roi	label	name	size	symmetry	laterality	x	y	z	neighbour	parent	overlap
-1	PUT	PUTAMEN	1033	0.88	0.032	-1.18	1.38	1.51	[2, 4, 8, 17, 18, 25, 36, 41, 46, 51]	2	1.0
-2	PINS_v	POSTERIOR_INSULA_ventral	420	0.66	0.024	-2.01	-7.26	-3.49	[1, 18, 24, 36, 46, 51, 64]	19	1.0
-3	CERVM	CEREBELLUM_VERMIS	360	0.8	0.003	0.2	-45.23	-8.42	[4, 6, 16, 25, 29, 35, 38, 48, 58]	30	1.0
-4	THAL	THALAMUS	805	0.9	-0.05	0.7	-18.24	6.0	[1, 3, 17, 25]	8	1.0
-5	CER7ab	CEREBELLUM_VIIIab	517	0.83	0.06	-1.59	-53.96	-51.08	[11, 15, 29, 44, 48]	13	1.0
-6	CER5	CEREBELLUM_I-V	747	0.91	0.0069999999999999	-0.4	-46.11	-23.17	[3, 16, 25, 28, 29, 35, 38, 48]	1	1.0
-7	ACcor_d	ANTERIOR_CINGULATE_CORTEX_dorsal	404	0.77	0.074	-0.29	20.72	28.48	[10, 21, 34, 39, 52, 60]	29	1.0
-8	CAUDNH_NACC	CAUDATE_NUCLEUS_HEAD_and_NUCLEUS_ACCUMBENS	285	0.83	0.046	-0.84	11.28	-11.21	[1, 17, 25, 34, 36, 41, 42]	10	1.0
-9	TP	TEMPORAL_POLE	753	0.73	0.254	-11.53	5.35	-40.93	[22, 24, 25, 36]	12	1.0
-10	DMPFcor_p	DORSOMEDIAL_PREFRONTAL_CORTEX_posterior	372	0.68	-0.032	0.02	30.65	40.42	[7, 34, 39, 52, 60, 61]	29	1.0
-11	R_CERCR2	CEREBELLUM_CRUSII_right	977	0.0	-1.0	32.67	-68.46	-38.41	[5, 12, 15, 29, 43, 48]	20	1.0
-12	CERCR1	CEREBELLUM_CRUSI	1133	0.76	0.098	-3.75	-79.71	-21.9	[11, 15, 28, 29, 30, 33, 38, 43, 44, 59]	7	1.0
-13	L_IFsul	left_INFERIOR_FRONTAL_SULCUS	702	0.0	1.0	-45.37	20.36	21.66	[14, 18, 50, 51, 53, 56, 63]	35	1.0
-14	MOTnet_vl	SOMATOMOTOR_NETWORK_ventrolateral	913	0.78	0.025	-2.15	-5.03	28.29	[13, 18, 23, 26, 46, 47, 51, 54, 56, 63]	3	1.0
-15	CER7b	CEREBELLUM_VIIb	816	0.66	-0.104	-0.1	-71.52	-52.09	[5, 11, 12, 29, 44, 48]	6	1.0
-16	FUSgyr_m	FUSIFORM_GYRUS_medial	426	0.62	0.1969999999999999	-5.12	-45.23	-11.99	[3, 6, 22, 25, 28, 29, 35, 38, 40, 58, 59]	30	1.0
-17	CAUDN	CAUDATE_NUCLEUS	638	0.84	0.011	-0.63	10.49	9.56	[1, 4, 8, 41, 42, 51]	8	1.0
-18	PINS_d	POSTERIOR_INSULA_dorsal	769	0.56	-0.181	6.45	-9.52	13.51	[1, 2, 13, 14, 23, 46, 47, 51, 54]	25	1.0
-19	PCcor_PRC	POSTERIOR_CINGULATE_CORTEX_and_PRECUNEUS	776	0.81	-0.048	0.7	-48.75	32.07	[21, 32, 35, 37, 57, 58]	4	1.0
-20	L_IPlob	left_INFERIOR_PARIETAL_LOBULE	724	0.0	1.0	-49.7	-56.98	37.77	[23, 46, 47, 57, 63, 64]	36	1.0
-21	PCcor_d	POSTERIOR_CINGULATE_CORTEX_dorsal	356	0.74	0.143	-0.9	-16.46	31.64	[7, 19, 32, 58, 60]	14	1.0
-22	ITgyr	INFERIOR_TEMPORAL_GYRUS	947	0.71	-0.172	8.61	-17.41	-27.84	[9, 16, 24, 25, 28, 36, 40, 64]	12	1.0
-23	SMgyr	SUPRAMARGINAL_GYRUS	779	0.54	0.2319999999999999	-13.94	-27.48	29.73	[14, 18, 20, 31, 46, 47, 49, 56, 63, 64]	33	0.926829268292683
-24	MTgyr_a	MIDDLE_TEMPORAL_GYRUS_anterior	936	0.75	0.021	-2.88	-4.38	-21.88	[2, 9, 22, 28, 36, 46, 51, 64]	31	1.0
-25	AMY_HIPP	AMYGDALA_and_HIPPOCAMPUS	1072	0.9	-0.009	-0.31	-19.0	-14.37	[1, 3, 4, 6, 8, 9, 16, 22, 28, 35, 36, 41, 58]	30	1.0
-26	FEF	FRONTAL_EYE_FIELD	725	0.7	-0.128	3.62	0.78	56.75	[14, 39, 45, 54, 56, 60, 61, 63]	21	1.0
-27	VISnet_md	VISUAL_NETWORK_mediodorsal	1023	0.81	0.012	0.08	-84.0	27.99	[30, 33, 35, 37, 43, 57, 58, 59]	11	1.0
-28	OCCTgyr_l	OCCIPITOTEMPORAL_GYRUS_lateral	952	0.64	0.265	-12.9	-44.83	-17.53	[6, 12, 16, 22, 24, 25, 29, 38, 40, 59, 64]	16	1.0
-29	CER6	CEREBELLUM_VI	1132	0.85	0.015	-0.36	-60.13	-27.13	[3, 5, 6, 11, 12, 15, 16, 28, 35, 38, 43, 44, 48, 59]	13	1.0
-30	PVISnet_m	POSTERIOR_VISUAL_NETWORK_medial	685	0.75	-0.188	6.2	-92.47	0.16	[12, 27, 38, 43, 59]	17	1.0
-31	R_IPlob	right_INFERIOR_PARIETAL_LOBULE	656	0.0	-1.0	50.44	-57.38	37.21	[23, 46, 49, 57, 64]	23	1.0
-32	IMsul	INFERIOR_MARGINAL_SULCUS	751	0.83	0.024	-0.42	-30.76	47.29	[19, 21, 37, 45, 47, 57, 60, 62]	14	1.0
-33	MVISnet_p	MEDIAL_VISUAL_NETWORK_posterior	377	0.72	-0.077	0.89	-85.79	6.56	[12, 27, 35, 38, 43, 58]	26	1.0
-34	PGACcor	PERIGENUAL_ANTERIOR_CINGULATE_CORTEX	644	0.74	0.019	-0.74	39.61	9.37	[7, 8, 10, 39, 41, 42, 52, 53, 55, 61]	18	1.0
-35	MVISnet_a	MEDIAL_VISUAL_NETWORK_anterior	1015	0.82	0.036	0.05	-62.44	3.99	[3, 6, 16, 19, 25, 27, 29, 33, 38, 43, 57, 58]	26	1.0
-36	AINS_v_PIsul	ANTERIOR_INSULA_ventral_and_PERI_INSULAR_SULCUS	1013	0.82	-0.031	0.14	12.11	-20.17	[1, 2, 8, 9, 22, 24, 25, 41, 46, 50, 51, 64]	19	1.0
-37	PRC_d	PRECUNEUS_dorsal	868	0.79	-0.039	0.09	-60.37	54.22	[19, 27, 32, 45, 57, 58, 62, 63]	32	1.0
-38	LVISnet_a	LATERAL_VISUAL_NETWORK_anterior	722	0.63	-0.249	6.54	-67.88	-9.04	[3, 6, 12, 16, 28, 29, 30, 33, 35, 40, 43, 59]	24	0.6135734072022161
-39	R_MFgyr_p	right_MIDDLE_FRONTAL_GYRUS_posterior	777	0.0	-1.0	27.31	26.74	42.92	[7, 10, 26, 34, 49, 52, 53, 54, 60, 61]	23	0.7438867438867439
-40	FUSgyr_l	FUSIFORM_GYRUS_lateral	951	0.6	-0.033	4.39	-60.32	-0.66	[16, 22, 28, 38, 43, 57, 59, 64]	16	1.0
-41	ORBcor	ORBITOFRONTAL_CORTEX	1126	0.8	0.064	-2.18	36.73	-19.52	[1, 8, 17, 25, 34, 36, 42, 50, 51, 53, 55]	10	1.0
-42	VMPFcor	VENTRAL_MEDIAL_PREFRONTAL_CORTEX	646	0.81	-0.008	0.14	51.14	-8.9	[8, 17, 34, 41, 52, 53, 55]	18	1.0
-43	PVISnet_l	POSTERIOR_VISUAL_NETWORK_lateral	1052	0.73	-0.049	-0.55	-91.57	1.76	[11, 12, 27, 29, 30, 33, 35, 38, 40, 57, 59]	17	1.0
-44	L_CERCR2	CEREBELLUM_CRUSII_left	940	0.0	0.999	-30.27	-72.08	-38.44	[5, 12, 15, 29, 48]	20	1.0
-45	MOTnet_m	SOMATOMOTOR_NETWORK_medial	1487	0.72	0.11	-2.84	-21.93	62.22	[26, 32, 37, 47, 56, 60, 62, 63]	15	1.0
-46	AUDnet	AUDITORY_NETWORK	1749	0.81	0.121	-6.01	-24.33	8.7	[1, 2, 14, 18, 20, 23, 24, 31, 36, 51, 64]	25	1.0
-47	PCsul	POSTCENTRAL_SULCUS	1255	0.67	-0.087	4.73	-32.52	49.62	[14, 18, 20, 23, 32, 45, 49, 56, 57, 62, 63]	33	1.0
-48	CER9	CEREBELLUM_IX	966	0.87	0.062	-0.57	-53.31	-45.54	[3, 5, 6, 11, 15, 29, 44]	5	1.0
-49	R_IPsul_PORB	right_INTRAPARIETAL_SULCUS_and_PARS_ORBITALIS	708	0.0	-1.0	43.54	-3.08	29.25	[23, 31, 39, 47, 50, 51, 53, 54, 55, 57, 62]	35	1.0
-50	VLPFcor	VENTROLATERAL_PREFRONTAL_CORTEX	1105	0.65	0.285	-12.66	32.73	-4.54	[13, 36, 41, 49, 51, 53, 54, 55]	34	1.0
-51	AINS_d	ANTERIOR_INSULA_dorsal	1317	0.68	-0.13	5.97	15.46	2.37	[1, 2, 13, 14, 17, 18, 24, 36, 41, 46, 49, 50, 53, 54]	28	1.0
-52	DMPFcor_a	DORSOMEDIAL_PREFRONTAL_CORTEX_anterior	1173	0.81	-0.06	1.39	56.87	18.08	[7, 10, 34, 39, 42, 53, 55, 61]	36	0.9616368286445012
-53	MFgyr_a	MIDDLE_FRONTAL_GYRUS_anterior	1206	0.49	0.461	-16.87	47.05	18.55	[13, 34, 39, 41, 42, 49, 50, 51, 52, 54, 55, 61, 63]	29	0.7736318407960199
-54	R_IFsul	right_INFERIOR_FRONTAL_SULCUS	977	0.0	-1.0	45.11	15.03	30.2	[14, 18, 26, 39, 49, 50, 51, 53, 56]	35	0.7185261003070624
-55	FP_l	FRONTAL_POLE_lateral	783	0.48	-0.374	11.97	57.22	-6.12	[34, 41, 42, 49, 50, 52, 53]	34	1.0
-56	MOTnet_dl	SOMATOMOTOR_NETWORK_dorsolateral	1026	0.65	-0.166	6.8	-17.21	54.68	[13, 14, 23, 26, 45, 47, 54, 63]	9	1.0
-57	DVIS	DORSAL_VISUAL_STREAM	1683	0.61	-0.175	6.11	-69.37	39.12	[19, 20, 27, 31, 32, 35, 37, 40, 43, 47, 49, 58, 59, 62, 63, 64]	22	1.0
-58	POsul	PARIETO_OCCIPITAL_SULCUS	1039	0.79	0.12	-1.96	-60.54	22.97	[3, 16, 19, 21, 25, 27, 33, 35, 37, 57, 63]	32	1.0
-59	LVISnet_p	LATERAL_VISUAL_NETWORK_posterior	1369	0.64	-0.085	2.94	-78.28	3.86	[12, 16, 27, 28, 29, 30, 38, 40, 43, 57]	24	0.933528122717312
-60	SMOTcor	PRE_SUPPLEMENTARY_MOTOR_CORTEX	1319	0.82	0.01	-0.46	4.63	57.03	[7, 10, 21, 26, 32, 39, 45, 61, 62]	21	1.0
-61	L_SFgyr_a	SUPERIOR_FRONTAL_GYRUS_anterior	1258	0.45	0.483	-11.46	34.66	46.54	[10, 26, 34, 39, 52, 53, 60, 63]	36	1.0
-62	SPlob	SUPERIOR_PARIETAL_LOBULE	1032	0.73	0.033	-1.6	-44.08	67.97	[32, 37, 45, 47, 49, 57, 60, 63]	27	0.8003875968992248
-63	L_IPsul_MFgyr_p	left_INTRAPARIETAL_SULCUS_and_MIDDLE_FRONTAL_GYRUS_posterior	1055	0.0	1.0	-39.19	-21.16	46.21	[13, 14, 20, 23, 26, 37, 45, 47, 53, 56, 57, 58, 61, 62]	35	0.4710900473933649
-64	MTgyr_p	MIDDLE_TEMPORAL_GYRUS_posterior	1568	0.78	-0.034	2.62	-42.23	4.33	[2, 20, 22, 23, 24, 28, 31, 36, 40, 46, 57]	31	1.0
