--- fmriprep_denoise/utils/data/custome_templateflow/tpl-MNI152NLin2009bSym/tpl-MNI152NLin2009bSym_res-03_atlas-MIST_desc-7_dseg.tsv
@@ -1,8 +1,0 @@
-roi	label	name	size	symmetry	laterality	x	y	z	neighbour	parent	overlap
-1	CER	CEREBELLUM	7228	0.96	0.014	-0.41	-64.36	-35.89	[2, 4, 6]		0
-2	LIMnet	MESOLIMBIC_NETWORK	5556	0.87	0.013	-0.6	1.39	-21.82	[1, 4, 5, 6, 7]		0
-3	MOTnet	SOMATOMOTOR_NETWORK	4780	0.89	0.004	-0.29	-20.89	54.05	[5, 6, 7]		0
-4	VISnet	VISUAL_NETWORK	6728	0.9	-0.047	1.4	-78.78	7.01	[1, 2, 5, 6]		0
-5	DMnet	DEFAULT_MODE_NETWORK	8986	0.78	0.111	-4.37	0.0	14.04	[2, 3, 4, 6, 7]		0
-6	FPnet_VISDN	FRONTO_PARIETAL_NETWORK_and_VISUAL_DOWNSTREAM	12206	0.78	-0.053	2.09	-13.48	22.88	[1, 2, 3, 4, 5, 7]		0
-7	VATTnet_SALnet_BG_THAL	VENTRAL_ATTENTION_NETWORK_and_SALIENCE_NETWORK_and_BASAL_GANGLIA_and_THALAMUS	12306	0.86	0.004	-0.29	-9.26	24.52	[2, 3, 5, 6]		0
